test that `on`- and `addEventListener` callbacks fire on passthrough
This change ensures that native window.XMLHttpRequest is restored
in passthrough tests, even after pretender's reference to
`_nativeXMLHttpRequest` is changed in the tests.

Also adds tests that listeners added with `addEventListener` are called
on a passed-through xhr. FakeXMLHttpRequest's `dispatchEvent` method
must be called with the event (which must have a `type` key) instead of
`(eventName, event)`.