Passthrough - fixes #121
Handler for passthrough should not be modified
so that it will match PASSTHROUGH in the check
for a passthrough request.

Also, make sure we don't overwrite the native
XMLHttpRequest function that we save. If the user
forgets to call shutdown(), the saved value was
getting overwritten with fakeRequest on the next
instantiation of Pretender.