Avoid setting `QUnit.module.exports = Pretender`.
When using `QUnit` along with `Pretender`, the UMD style that we were previously using would assign an `.exports` property to the `QUnit.module` method (which is present in QUnit 1.x as `self.module` also). This changes to check that we are actually in a node-like environment (where `exports === module.exports`) before setting `module.exports`.